fix: Allow TemporaryDirectory to be disposed multiple times

+++ src/Utilities/IO/TemporaryDirectory.cs
@@ -37,16 +37,19 @@ namespace Grynwald.Utilities.IO
         /// </summary>
         public void Dispose()
         {
-            // remove read-only flag from all files
-            var files = new DirectoryInfo(FullName).GetFiles("*", SearchOption.AllDirectories);
-            foreach (var file in files)
-            {
-                file.IsReadOnly = false;
-            }
+            if (!Directory.Exists(FullName))
+                return;
 
-            // recursively delete the directory
             try
             {
+                // remove read-only flag from all files
+                var files = new DirectoryInfo(FullName).GetFiles("*", SearchOption.AllDirectories);
+                foreach (var file in files)
+                {
+                    file.IsReadOnly = false;
+                }
+
+                // recursively delete the directory
                 Directory.Delete(FullName, recursive: true);
             }
             catch (UnauthorizedAccessException)
